Move ThreadPoolWorker into cvat.apps.django_rq_ext
