+++ cvat/apps/django_rq_ext/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DjangoRqExtConfig(AppConfig):
+    name = "cvat.apps.django_rq_ext"
