+++ cvat/apps/django_rq_ext/sandbox/tasks.py
@@ -1,0 +1,10 @@
+import time
+
+
+def send_webhook_task(url: str, method: str = "GET", payload: dict | None = None) -> dict:
+    time.sleep(10)
+    return {
+        "status_code": 20,
+        "url": url,
+        "method": method,
+    }
