--- cvat/apps/rq_playground/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class RqPlaygroundConfig(AppConfig):
-    name = "cvat.apps.rq_playground"
